Discord_Bot/requirements.txt
discord.py==2.3.2
playwright==1.42.0
